Clarify JsPrettyNamer algorithm
Introduced a changeNames() method to clarify the dataflow between
parent and child during recursion. Also, changed the return type of
JsScope.getAllNames() for convenience.

There should be no change in behavior.

(This is in preparation for adding a way to reduce the number of
globals.)

Change-Id: I0c30b8be599dad8f6b881c6bd24237241031e1f5
Review-Link: https://gwt-review.googlesource.com/#/c/5220/