Removes dead code from DOMImplStandard
There exists a disposeEvent method in DOMImplStandard that is not
used in any place other than DOMImplMozilla which just overrides.
It looks like real intention was to override disposeEventSystem
and remove the capture handler.

This patch removes the disposeEvent method and fixes up the logic
for removing the capture handler.

Change-Id: I96aa7a9b23178329beecda466e65cb661ed8a55d
Review-Link: https://gwt-review.googlesource.com/#/c/5221/