Introduces GWT.reportUncaughtException.
This patch adds GWT.reportUncaughtException as a replacement of the
earlier introduced GWT.maybeReportUncaughtException.

Unlike the previous version, as the name suggests, new one guarantees
that the exception is not swallowed and reported back to the browser
even there is no uncaught exception handler.

Bug: issue 6833
Change-Id: Ica7e86f73807f637535c7fd73b8256189385f9e7
Review-Link: https://gwt-review.googlesource.com/#/c/5416/