Adds the RuntimeRebinder registry.
To perform runtime rebinding one needs a central runtime registry
of rebind rules and some logic to find and apply the matching
rule when rebinds are requested. RuntimeRebinder is just that
central registry and rebinder.

Runtime rebinding is needed for separate compilation since static
compile time rebinding is only possible when in possession of complete
global knowledge and thus can only be done in monolithic compiles.

Change-Id: I97f2233301a8a67783e54dfe100217182a647f78