Change UserAgentAsserter to use UCE instead of a dialog
When a dialog is used, end users are facing with an obscure message and
the developers will not even be aware of the problem.
With UCE, the error will be piped through exception handling mechanism
that is already setup for the environment (e.g. remote logging in
production).

Bug: issue 8079
Change-Id: Ibac3460eedd140b604ef31ec5d4ce693306b801f
Review-Link: https://gwt-review.googlesource.com/#/c/5930/