adds a LibraryCompiler entry point.
The LibraryCompiler entry point is not yet ready to use and depends
on changes that are still being reviewed.

Reading/writing libraries wasn't keeping track of generated resources
and so separate compilation was failing to provide all generated
resources to the final link process. Fixed.

Generated resources were keeping a reference to the generator that
created them, but this information was never used and was making it
difficult to read/write generated resources in libraries. Removed.

StandardLinkerContext was reading public resources directly out of the
module instance, which is only safe when there are no libraries which
might also be providing public resources. Changed it to receive and
operate on a ResourceOracle instance and updated all
StandardLinkerContext instantiation to pass in the instance found by
querying the compilerContext for its public ResourceOracle (since that
oracle will be correct for both monolithic and separate compiles). This
new need to provide a public ResourceOracle instance necessitated some
new parameters to functions in the Link.link() call tree.

Added library compilation specific options (link, input libraries,
output library) to CompileTaskOptions and made sure registered
matching arg handlers in the new LibraryCompiler entry point.

Caught and fixed a couple of entry points which were constructing
compilerContext instances and moduleDef instances but were failing to
store the module reference in the compilerContext. In the past it
didn't matter but now that compilerContext.getPublicResourceOracle()
is always used instead of module.getPublicResource() it started to
matter.

Flipped the name and meaning of the
gatherLibrariesForGenerator(..., boolean generatorWasRun) boolean to
make the method body more readable.

Lowered compression level and added stream buffering to improve
ZipLibraryWriter write speed.

Change-Id: I881c8d2f3fbe88e2b10b4c9a009b16d486dbd350
Review-Link: https://gwt-review.googlesource.com/#/c/6023/