Revert "Makes UnitCache keys predictable (removes path prefix)."
After a lot of digging I've found that this is not safe.

There is an undocumented requirement that UnitCache implementations
must make it safe to reuse a single UnitCache instance across multiple
compiles of different applications within the same process.

If one of those applications uses foo/Bar.java and the other uses
super/foo/Bar.java, this change will cause one of the compiles to get
the wrong version.

This reverts commit 37ddbe4b38c10fb3df2ced029eed2f0b75cbd7f0.

Change-Id: Id63695db9477a3a1d80f76999d09911daa70ef0d