Define trampolines at actual JSO class rather than in JavaScriptObject.
Trampoline methods where declared in JavaScriptObject instead of the
actual JSO class. In modular compilation JavaScriptObject code will
only be generated in the defining module, hence separate compilation
would need to be aware to regenerate JavaScriptObject.

Also removed some redundant generated code in the String specific
trampolines.

In the current scheme still JSOs need to be declared in the SAME
module as the interfaces they implement when compiling modularly.

TODO:
  - Lift the above restriction.

Change-Id: Id563afe7650b099f6cfafbb814e24dac1ed2dbc4