Fix daylight time savings bug.
Creating a Date(year, month, day, hour, minute, sec, miliseconds) is
ambiguous if the time falls in the hour after the time was pulled back.

In Java semantics the date created will refer to the latter time,
whereas GWT was relying on the browser to disambiguate, and a recent
release of Firefox handles this case differently.

Change-Id: Ic097bb3bc38862b363dbb377d6c5f9f2f045028f