Supports overrides with co-variant return types in RequestFactory, and generics in proxies.
Generics in proxies can lead to overrides with co-variant return types
(e.g. if you override a method to set a @PropertyName on it), but also has
its own set of issues (type variables in getter's return type –used as the
property type– have to be "resolved" against the actual interface
parameterization).

Bug: issue 5926
Change-Id: Iba6384a93b866934c6e744643d79f9063e36bb2d