Introduces copy-less iterator for InternalJsStringMap
Refactors hashmap to provide iteration capability without the costly
'iterate and construct all entries' approach for the modern browsers.
The should provide huge memory savings and speed-up for big maps.

As the first step, this is implemented only for InternalJsStringMap
iterator. A follow up patch will introduce the similar approach to
InternalHashCodeMap.

Change-Id: If1c0acdbd09f0e62321b847a009ae6f7ebc59834