Enables logging by default for superdevmode
This change introduce a very light weight java.util.Logger based
logging to console and enables it by default for super-dev mode.

To do that it introduces non-public LoggerConfigurator API that is
much simpler than LoggerConfiguration but also enables to have more
complex logging configuration for apps (e.g. per package configuration).
But as it is not the goal of this patch it is kept private until we
decide how to move forward.

Change-Id: Iea106eebbd241db233971eef675bd5ad26d4c116