Make History behave with GwtMockito, IE10 & IE11 and old FF.
Since GwtMockito can not mock static native methods, moved JSNI code
into impl classes.
Removed the pushstate and popstate implementation since IE10 & IE11
do not fire a popstate event if a hashchange is triggered due to
click on an anchor in the page. Instead using the hashchange event for
all browsers.
Old versions of firefox require to add a third parameter for event
registration (the boolean for bubbling). Added this to make History
work in FF < 6.
Fix HistoryTest#testClickLink for IE8 & IE9.

Change-Id: I36143945224710cd067479ebd780b2adbf3f5c58