Fix handling of fields whose unitialized values can be seen.
Uninitialized values for intance fields that are assigned in the
constructor might be observed under certain circumstances.

Some optimizations assumed that if a field is ONLY iniitialized in
the constructor to a value literal, its uninitilized values could not
be seen, but that is not the case.

- Added test cases to demonstrate the issues.
- Fixed Finalizer.

This patch will affect code size since the optimizations performed
were not sound. A follow up patch will reintroduce the fixed
optimizations.

Change-Id: I44d4c5e83b879faa042c69989ea6b1a95400159d