Removes size book keeping from ArrayList.
ArrayList used to keep track of its own size because of ensureCapacity.
Changed the implementation to ignore the capacity change requests so
we can use array.length instead of tracking the size manually.

This both simplifies the implementation but will also let us compile
ArrayList down to a simple array after we make the collection
bounds checks optional.

Change-Id: Ic495cd296e194739b1c92d5ec3e57b3051136524