Extends staleness marking incase of unstable Generator output.
Previously if Generator A contained a nested type that referenced some
type Foo on disk. When Foo changed we would rerun Generator A and
reprocess just the known-referencing nested type.

But if the Generator's output was unstable then some of the names of
anonymous types can change and invalidate references between the types
that were and were not reprocessed.

We now assume all types for the rerun generator may have modified and
thus reprocess all of them.

This may be a noticeable performance decrease for Generators that run
once with a lot of output, like GWTRPC. There are more complicated
approaches that are faster, so this can be revisited.

Bug: issue 8880.
Change-Id: I94e63d79cbdf541ce519e4139d6a23228eae98a9