Fixed ICE in EqualityNormalizer due to bug in Pruner.
The Pruner would leave the tree in an inconsistent state if an
array of primitive type was determined to be uninstantiable.
In such case an array reference if the form "c.a[x]" where "c.a"
was typed as "int[]" would be replaced by "c.nullField[x]" where
"c.nullField" was typed as JNullType. That makes an expression like
"c.a[x] == 1" be illtyped (in a way that was not expected by
EqualityNormalizer).

Bug: issue 2478.
Change-Id: Iec951aa729910ddc79293c841845c75ed11105e3