Fixes incorrect linked type order changes after recompile.
Normally we sort linked output alphabetically but with super classes
before sub classes.

But if a super class was marked stale, and we threw out type index
information for it, we were also throwing out the sub class -> super
class references, and not rebuilding them (since none of the subclasses
are stale).

Changed the way the index is pruned to no longer trim information that
won't be rebuilt by processing the stale type.

Change-Id: Ifc084a703ebd2a8a72a1bd8e1f640e25f00ad7e3